# perf!: reduce crc16 pointer dereference