--- a/decoder/decoder_test.go
+++ b/decoder/decoder_test.go
@@ -385,7 +385,7 @@ func TestCheckIntegrity(t *testing.T) {
 					DataType:        proto.DataTypeFIT,
 				}
 				b, _ := h.MarshalBinary()
-				crc := crc16.New(crc16.MakeFitTable())
+				crc := crc16.New(nil)
 				crc.Write(b[:12])
 				binary.LittleEndian.PutUint16(b[12:14], crc.Sum16())
 				return bytes.NewReader(b)
@@ -539,7 +539,7 @@ func createFitForTest() (proto.FIT, []byte) {
 	bytesbuffer.Write(b)
 
 	// Marshal and calculate data size and crc checksum
-	crc16checker := crc16.New(crc16.MakeFitTable())
+	crc16checker := crc16.New(nil)
 	for i := range fit.Messages {
 		mesg := fit.Messages[i]
 		mesgDef := proto.CreateMessageDefinition(&mesg)