--- a/encoder/encoder_test.go
+++ b/encoder/encoder_test.go
@@ -477,7 +477,7 @@ func TestEncodeHeader(t *testing.T) {
 
 				binary.LittleEndian.PutUint16(b[2:4], profile.Version)
 
-				crc := crc16.New(crc16.MakeFitTable())
+				crc := crc16.New(nil)
 				crc.Write(b[:12])
 				binary.LittleEndian.PutUint16(b[12:14], crc.Sum16())