--- a/kit/hash/crc16/crc16_test.go
+++ b/kit/hash/crc16/crc16_test.go
@@ -5,42 +5,55 @@
 package crc16_test
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/muktihari/fit/kit/hash/crc16"
 )
 
 func TestCRC16(t *testing.T) {
-	b := []byte{14, 32, 84, 8, 214, 204, 9, 0, 46, 70, 73, 84} // example from some fit header.
-	crc := uint16(12856)                                       // should match this checksum.
-
-	c16 := crc16.New(nil)
-
-	if c16.BlockSize() != 1 {
-		t.Fatalf("blocksize mismatch")
+	tt := []struct {
+		name  string
+		table *crc16.Table
+	}{
+		{name: "nil table", table: nil},
+		{name: "crc16.MakeFITTable()", table: crc16.MakeFITTable()},
 	}
 
-	if c16.Size() != 2 {
-		t.Fatalf("size mismatch")
-	}
+	for i, tc := range tt {
+		t.Run(fmt.Sprintf("[%d] %s", i, tc.name), func(t *testing.T) {
+			b := []byte{14, 32, 84, 8, 214, 204, 9, 0, 46, 70, 73, 84} // example from some fit header.
+			crc := uint16(12856)                                       // should match this checksum.
 
-	_, _ = c16.Write(b)
-	if c16.Sum16() != crc {
-		t.Fatalf("expected: %v, got: %v", crc, c16.Sum16())
-	}
+			c16 := crc16.New(tc.table)
 
-	sum := c16.Sum([]byte{10})
-	expect := []byte{10, 50, 56}
-	for i := range sum {
-		if sum[i] != expect[i] {
-			t.Fatalf("expected: %v, got: %v", expect[i], sum[i])
-		}
-	}
+			if c16.BlockSize() != 1 {
+				t.Fatalf("blocksize mismatch")
+			}
 
-	c16.Reset()
+			if c16.Size() != 2 {
+				t.Fatalf("size mismatch")
+			}
 
-	if c16.Sum16() != 0 {
-		t.Fatalf("expected 0 after reset, got: %v", c16.Sum16())
+			_, _ = c16.Write(b)
+			if c16.Sum16() != crc {
+				t.Fatalf("expected: %v, got: %v", crc, c16.Sum16())
+			}
+
+			sum := c16.Sum([]byte{10})
+			expect := []byte{10, 50, 56}
+			for i := range sum {
+				if sum[i] != expect[i] {
+					t.Fatalf("expected: %v, got: %v", expect[i], sum[i])
+				}
+			}
+
+			c16.Reset()
+
+			if c16.Sum16() != 0 {
+				t.Fatalf("expected 0 after reset, got: %v", c16.Sum16())
+			}
+		})
 	}
 }