BREAKING CHANGES:
- `MakeFitTable` renamed to `MakeFITTable`

Benchmark:
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/kit/hash/crc16
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
        │   old.txt   │               new.txt               │
        │   sec/op    │   sec/op     vs base                │
Write-4   27.82µ ± 0%   21.55µ ± 0%  -22.53% (p=0.000 n=10)

        │  old.txt   │            new.txt             │
        │    B/op    │    B/op     vs base            │
Write-4   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

        │  old.txt   │            new.txt             │
        │ allocs/op  │ allocs/op   vs base            │
Write-4   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

```

close #180